Move contents of `dynamic_operator.hpp` into `pipeline.hpp`
