+++ libvast/builtins/operators/pass.cpp
@@ -6,7 +6,7 @@
 // SPDX-FileCopyrightText: (c) 2021 The VAST Contributors
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include "vast/dynamic_operator.hpp"
+#include "vast/pipeline.hpp"
 
 #include <vast/concept/parseable/string/char_class.hpp>
 #include <vast/concept/parseable/vast/pipeline.hpp>
@@ -45,7 +45,7 @@ private:
 // Does nothing with the input.
 class pass_operator2 final : public crtp_operator<pass_operator2> {
 public:
-  template <input_element T>
+  template <operator_input_element T>
   auto operator()(T x) const -> T {
     return x;
   }
